# api/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ListReportsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ReportDTO'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationMetadata'
    type: object
...
  dto.NearbyReportsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ReportWithDistance'
        type: array
    type: object
...
  dto.PaginationMetadata:
    properties:
      has_more:
        type: boolean
      has_previous:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.ReportCreate:
    properties:
      address:
        type: string
      description:
        type: string
      image_url:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      municipality:
        type: string
      neighborhood:
        type: string
      province:
        type: string
      risk_topic_id:
        type: string
      risk_type_id:
        type: string
      user_id:
        type: string
    type: object
  dto.ReportDTO:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      expires_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      municipality:
        type: string
      neighborhood:
        type: string
      province:
        type: string
      rejection_count:
        type: integer
      resolved_at:
        type: string
      reviewed_by:
        type: string
      risk_topic_icon_url:
        type: string
      risk_topic_id:
        type: string
      risk_topic_name:
        type: string
      risk_type_icon_url:
        type: string
      risk_type_id:
        type: string
      risk_type_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      verification_count:
        type: integer
    type: object
  dto.ReportWithDistance:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      distance:
        description: Distance in meters
        type: number
      expires_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      municipality:
        type: string
      neighborhood:
        type: string
      province:
        type: string
      rejection_count:
        type: integer
      resolved_at:
        type: string
      reviewed_by:
        type: string
      risk_topic_icon_url:
        type: string
      risk_topic_id:
        type: string
      risk_topic_name:
        type: string
      risk_type_icon_url:
        type: string
      risk_type_id:
        type: string
      risk_type_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      verification_count:
        type: integer
    type: object
  dto.ResolveReportRequest:
    properties:
      moderator_id:
        type: string
    required:
    - moderator_id
    type: object
...
  dto.UpdateReportLocationRequest:
    properties:
      address:
        type: string
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      municipality:
        type: string
      neighborhood:
        type: string
      province:
        type: string
    required:
    - latitude
    - longitude
    type: object
  dto.UpdateReportLocationResponse:
    properties:
      id:
        type: string
      message:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.VerifyReportRequest:
    properties:
      moderator_id:
        type: string
    required:
    - moderator_id
    type: object
  dto.VoteReportRequest:
    properties:
      vote_type:
        enum:
        - upvote
        - downvote
        type: string
    required:
    - vote_type
    type: object
  dto.VoteReportResponse:
    properties:
      rejection_count:
        type: integer
      report_id:
        type: string
      verification_count:
        type: integer
      vote_type:
        type: string
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: integer
          message: {}
        type: object
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@riskplace.ao
    name: API Support
    url: http://www.riskplace.ao
  description: |-
    This is the API documentation for the Risk Place Angola application.

    ## Environments
    - **Development**: https://risk-place-angola-904a.onrender.com
    - **Local**: http://localhost:8000

    ## WebSocket Integration Guide
    For detailed instructions on integrating WebSocket in mobile applications, please refer to our [WebSocket Notification Guide](https://github.com/risk-place-angola/backend-risk-place/blob/develop/docs/MOBILE_WEBSOCKET_INTEGRATION.md)
  title: Risk Place Angola API
  version: 1.0.0
paths:
...
  /reports:
    get:
      consumes:
      - application/json
      description: List all reports in the system with pagination and filters
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by status (pending, verified, resolved)
        in: query
        name: status
        type: string
      - description: 'Sort field (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort order (asc, desc) (default: desc)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all reports with pagination
      tags:
      - reports
    post:
      consumes:
      - application/json
      description: Create a new report
      parameters:
      - description: Report to create
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/dto.ReportCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReportDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new report
      tags:
      - reports
  /reports/{id}/location:
    put:
      consumes:
      - application/json
      description: Update the geographic location of a report (used when user drags
        marker on map)
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: New location data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReportLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateReportLocationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update report location
      tags:
      - reports
  /reports/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Resolve a report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Resolution data
        in: body
        name: resolve
        required: true
        schema:
          $ref: '#/definitions/dto.ResolveReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve a report
      tags:
      - reports
  /reports/{id}/verify:
    post:
      consumes:
      - application/json
      description: Verify a report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Verification data
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify a report
      tags:
      - reports
  /reports/{id}/vote:
    post:
      consumes:
      - application/json
      description: Upvote or downvote a report to verify its authenticity
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Device ID for anonymous users
        in: header
        name: X-Device-Id
        type: string
      - description: Vote data
        in: body
        name: vote
        required: true
        schema:
          $ref: '#/definitions/dto.VoteReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.VoteReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - OptionalAuth: []
      summary: Vote on a report
      tags:
      - reports
  /reports/nearby:
    get:
      consumes:
      - application/json
      description: List reports near the specified location with calculated distance
      parameters:
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: number
      - description: Radius in meters
        in: query
        name: radius
        required: true
        type: number
      - description: 'Maximum number of results (default: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NearbyReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List nearby reports with distance
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DeviceID:
    description: Unique device identifier for tracking and analytics.
    in: header
    name: X-Device-ID
    type: apiKey
swagger: "2.0"